Update and rename BankAccountManagement.java to OOP/BankAccountManagement.java
Changed the header, improved the indentation.
