--- BankAccountManagement.java
@@ -1,111 +1,0 @@
-import java.util.Scanner;
-
-/**
- * Bank Account Management System
- * Demonstrates basic OOP concepts like classes, objects, and methods.
- * Author: [Sreenivasulu-03]
- * Date: [2025-10-05]
- */
-
-class BankAccount {
-    private String accountNumber;
-    private String accountHolderName;
-    private double balance;
-
-    // Constructor to initialize account details
-    public BankAccount(String accountNumber, String accountHolderName, double initialBalance) {
-        this.accountNumber = accountNumber;
-        this.accountHolderName = accountHolderName;
-        this.balance = initialBalance;
-    }
-
-    // Deposit method
-    public void deposit(double amount) {
-        if (amount > 0) {
-            balance += amount;
-            System.out.println("Successfully deposited: ₹" + amount);
-        } else {
-            System.out.println("Deposit amount must be positive!");
-        }
-    }
-
-    // Withdraw method
-    public void withdraw(double amount) {
-        if (amount > 0 && amount <= balance) {
-            balance -= amount;
-            System.out.println(" Successfully withdrew: ₹" + amount);
-        } else if (amount > balance) {
-            System.out.println(" Insufficient balance!");
-        } else {
-            System.out.println(" Invalid withdrawal amount!");
-        }
-    }
-
-    // Display current balance
-    public void checkBalance() {
-        System.out.println(" Current balance: ₹" + balance);
-    }
-
-    // Display account details
-    public void displayAccountDetails() {
-        System.out.println("\n--- Account Details ---");
-        System.out.println("Account Holder: " + accountHolderName);
-        System.out.println("Account Number: " + accountNumber);
-        System.out.println("Balance: ₹" + balance);
-    }
-}
-
-public class BankAccountManagement {
-    public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-
-        // Take account details from user
-        System.out.print("Enter Account Holder Name: ");
-        String name = sc.nextLine();
-        System.out.print("Enter Account Number: ");
-        String number = sc.nextLine();
-        System.out.print("Enter Initial Balance: ₹");
-        double balance = sc.nextDouble();
-
-        BankAccount account = new BankAccount(number, name, balance);
-        int choice;
-
-        // Menu-driven system
-        do {
-            System.out.println("\n===== Bank Menu =====");
-            System.out.println("1. Deposit");
-            System.out.println("2. Withdraw");
-            System.out.println("3. Check Balance");
-            System.out.println("4. Display Account Details");
-            System.out.println("5. Exit");
-            System.out.print("Enter your choice: ");
-            choice = sc.nextInt();
-
-            switch (choice) {
-                case 1:
-                    System.out.print("Enter amount to deposit: ₹");
-                    double depositAmount = sc.nextDouble();
-                    account.deposit(depositAmount);
-                    break;
-                case 2:
-                    System.out.print("Enter amount to withdraw: ₹");
-                    double withdrawAmount = sc.nextDouble();
-                    account.withdraw(withdrawAmount);
-                    break;
-                case 3:
-                    account.checkBalance();
-                    break;
-                case 4:
-                    account.displayAccountDetails();
-                    break;
-                case 5:
-                    System.out.println(" Thank you for using the Bank Account Management System!");
-                    break;
-                default:
-                    System.out.println(" Invalid choice! Please try again.");
-            }
-        } while (choice != 5);
-
-        sc.close();
-    }
-}
